Make style in in {src/*/,}CMakeLists.txt more consistent
This commit makes the formatting style of CMakeLists.txt in the
root directory and CMakeLists.txt in src/ subdirectories more
consistent.  The new style is the style that has already been
most commonly used in these files.

A consistent style (also in CMake files) looks professional, is thus
more appealing e.g., to new developers, and makes it easier to deduce
a style for code additions.

Details:
 - indentation (also for continuation) is 4 spaces
 - spaces around parentheses are removed
 - CMake commands are lower-cased
 - consecutive mergeable commands are merged
   (e.g., set_target_properties() for the same target)

While at it, this commit applies minor fixes to the comments.